+++ src/services/openai.ts
@@ -1,6 +1,9 @@
 import { OpenAI } from "openai";
 import { getGlobalConfig } from "../utils/config";
 import { ProxyAgent, fetch } from 'undici'
+import { setSessionState, getSessionState } from "../utils/sessionState";
+import { logEvent } from "../services/statsig";
+
 
 export async function getCompletion(
   type: 'large' | 'small', 
@@ -11,6 +14,56 @@ export async function getCompletion(
   const baseURL = type === 'large' ? config.largeModelBaseURL : config.smallModelBaseURL
   const proxy = config.proxy ? new ProxyAgent(config.proxy) : undefined
   
+  const toolDescriptions = {}
+  
+  opts = structuredClone(opts)
+
+  if (getSessionState('modelErrors')[`${baseURL}:${opts.model}:1024`]) {
+    for(const tool of opts.tools) {
+      if(tool.function.description.length <= 1024) {
+        continue
+      }
+      let str = ''
+      let remainder = ''
+      for(let line of tool.function.description.split('\n')) {
+        if(str.length + line.length < 1024) {
+          str += line + '\n'
+        } else {
+          remainder += line + '\n'
+        }
+      }
+      logEvent('truncated_tool_description', {
+        name: tool.function.name,
+        original_length: tool.function.description.length,
+        truncated_length: str.length,
+        remainder_length: remainder.length,
+      })
+      tool.function.description = str
+      toolDescriptions[tool.function.name] = remainder
+    }
+    if(Object.keys(toolDescriptions).length > 0) {
+      let content = '<additional-tool-usage-instructions>\n\n'
+      for(const [name, description] of Object.entries(toolDescriptions)) {
+        content += `<${name}>\n${description}\n</${name}>\n\n`
+      }
+      content += '</additional-tool-usage-instructions>'
+
+      for(let i = opts.messages.length - 1; i >= 0; i--) {
+        if(opts.messages[i].role === 'system') {
+          opts.messages.splice(i + 1, 0, {
+            role: 'system',
+            content
+          })
+          break
+        }
+      }
+    }
+  } else if (getSessionState('modelErrors')[`${baseURL}:${opts.model}:max_completion_tokens`]) {
+    opts.max_completion_tokens = opts.max_tokens
+    delete opts.max_tokens
+  }
+
+
   if (opts.stream) {
     const response = await fetch(`${baseURL}/chat/completions`, {
       method: 'POST',
@@ -24,6 +77,17 @@ export async function getCompletion(
     
     if (!response.ok) {
       const error = await response.json() as { error?: { message: string } }
+      if (error.error?.message?.indexOf('Expected a string with maximum length 1024') !== -1) {
+        setSessionState('modelErrors', {
+          [`${baseURL}:${opts.model}:1024`]: error.error?.message
+        })
+        return getCompletion(type, opts)
+      } else if (error.error?.message?.indexOf("Use 'max_completion_tokens'") !== -1) {
+        setSessionState('modelErrors', {
+          [`${baseURL}:${opts.model}:max_completion_tokens`]: error.error?.message
+        })
+        return getCompletion(type, opts)
+      }
       throw new Error(`API request failed: ${error.error?.message || JSON.stringify(error)}`)
     }